requirements.txt
streamlit==1.32.0
pandas==2.2.1
openpyxl==3.1.2
plotly==5.19.0
openai==1.13.3
chardet==5.2.0
python-dotenv
pygwalker